iniciado layout
